# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tracks:
    post:
      operationId: createTrack
      tags: [tracks]
      summary: Upload a new track (artists only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackBody"
      responses:
        "201":
          description: Track created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Only artists can upload tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getTracks
      tags: [tracks]
      summary: List all tracks
      parameters:
        - in: query
          name: genre
          schema:
            type: string
          required: false
          description: Filter by genre
        - in: query
          name: limit
          schema:
            type: integer
          required: false
          description: Max number of results
        - in: query
          name: university
          schema:
            type: string
          required: false
          description: Filter by university
        - in: query
          name: artistId
          schema:
            type: string
          required: false
          description: Filter by artist user ID
      responses:
        "200":
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
  /tracks/trending:
    get:
      operationId: getTrendingTracks
      tags: [tracks]
      summary: Get trending tracks sorted by play count
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          required: false
          description: Max number of results (default 10)
      responses:
        "200":
          description: Trending tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
  /tracks/most-liked:
    get:
      operationId: getMostLikedTracks
      tags: [tracks]
      summary: Get tracks sorted by total like count (popularity leaderboard)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          required: false
          description: Max number of results (default 50)
      responses:
        "200":
          description: Most liked tracks sorted by descending like count
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
  /tracks/{id}:
    get:
      operationId: getTrackById
      tags: [tracks]
      summary: Get a single track by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Track detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrack
      tags: [tracks]
      summary: Update a track (owner only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackBody"
      responses:
        "200":
          description: Updated track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the track owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrack
      tags: [tracks]
      summary: Delete a track (owner only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "204":
          description: Track deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the track owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tracks/liked:
    get:
      operationId: getLikedTrackIds
      tags: [tracks]
      summary: Get liked track IDs for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of liked track IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /tracks/library:
    get:
      operationId: getLibraryTrackIds
      tags: [tracks]
      summary: Get library (saved) track IDs for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of library track IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /tracks/{id}/like:
    post:
      operationId: likeTrack
      tags: [tracks]
      summary: Like or unlike a track (persisted per-user when authenticated)
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeBody"
      responses:
        "200":
          description: Updated like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tracks/{id}/play:
    post:
      operationId: recordPlay
      tags: [tracks]
      summary: Increment the play count for a track
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Updated play count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tracks/{id}/library:
    post:
      operationId: saveToLibrary
      tags: [tracks]
      summary: Save or unsave a track from the library
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryBody"
      responses:
        "200":
          description: Updated library status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LikeBody:
      type: object
      properties:
        liked:
          type: boolean
    LikeResponse:
      type: object
      properties:
        trackId:
          type: string
        likes:
          type: integer
        liked:
          type: boolean
      required:
        - trackId
        - likes
        - liked
    LibraryBody:
      type: object
      properties:
        saved:
          type: boolean
    LibraryResponse:
      type: object
      properties:
        trackId:
          type: string
        saved:
          type: boolean
      required:
        - trackId
        - saved
...
    Track:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        artistId:
          type: string
        genre:
          type: string
        duration:
          type: string
        durationSeconds:
          type: integer
        coverColor:
          type: string
        audioUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        playCount:
          type: integer
        likes:
          type: integer
        university:
          type: ["string", "null"]
      required:
        - id
        - title
        - artist
        - artistId
        - genre
        - duration
        - durationSeconds
        - coverColor
        - audioUrl
        - coverUrl
        - playCount
        - likes
        - university
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    PlayResponse:
      type: object
      properties:
        trackId:
          type: string
        playCount:
          type: integer
      required:
        - trackId
        - playCount
    CreateTrackBody:
      type: object
      properties:
        title:
          type: string
        genre:
          type: string
        audioUrl:
          type: string
        coverUrl:
          type: ["string", "null"]
        duration:
          type: string
        durationSeconds:
          type: integer
      required:
        - title
        - genre
        - audioUrl
    UpdateTrackBody:
      type: object
      properties:
        title:
          type: string
        genre:
          type: string
        coverUrl:
          type: ["string", "null"]
...